// src/models/Subscription.ts
export interface LineItem extends BaseModel {
	readonly subscription_id: string;
	readonly customer_id: string;
	readonly plan_id: string;
	readonly price_id: string;
	readonly meter_id: string;
	readonly environment_id: string;
	readonly tenant_id: string;
	readonly display_name: string;
	readonly plan_display_name: string;
	readonly meter_display_name: string;
	readonly price_type: string;
	readonly billing_period: string;
	readonly currency: string;
	readonly quantity: number;
	readonly start_date: string;
	readonly end_date: string;
	readonly metadata: Metadata;
	readonly price?: Price;
	readonly subscription_phase_id?: string;
}
// ...
export enum SUBSCRIPTION_STATUS {
	ACTIVE = 'active',
	PAUSED = 'paused',
	CANCELLED = 'cancelled',
	INCOMPLETE = 'incomplete',
	INCOMPLETE_EXPIRED = 'incomplete_expired',
	PAST_DUE = 'past_due',
	TRIALING = 'trialing',
	UNPAID = 'unpaid',
	DRAFT = 'draft',
}
// ...
export enum SUBSCRIPTION_CANCELLATION_TYPE {
	IMMEDIATE = 'immediate',
	END_OF_PERIOD = 'end_of_period',
}